Upload MovieApp in Clean Archtecture

+++ lib/features/movie/data/dtos/movie_details_dto.dart
@@ -36,7 +36,7 @@ extension MovieDetailsDto on MovieDetailsEntity {
       releaseDate: json['release_date'],
       title: json['title'],
       video: json['video'],
-      voteAverage: json['vote_average'],
+      voteAverage: json['vote_average'].toDouble(),
       voteCount: json['vote_count'],
     );
   }
